crypto/symcipher: bind {enc,dec}ryption instances to block cipher mode
Currently, the SymBlockCipher{Encryption,Decryption}Instance would internally
store block cipher implementation instances capable of both, encryption and
decryption. This is so, because they can be used with any block cipher mode to
be selected on a per request basis and therefore it cannot be known upfront
whether the block cipher will be used in the encryption or decryption direction.
However, it is common that a given SymBlockCipher{Encryption,Decryption}Instance
will only ever be used with a single unique block cipher mode and in such
scenarios it is wasteful to prepare and store the additional block cipher
instance state for the unneeded direction.

Bind encryption and decryption instances to the block cipher mode at their
instantiation time, so that it is known ahead in which direction the underlying
block cipher implementation will be used.

Rename SymBlockCipherEncryptionInstance to SymBlockCipherModeEncryptionInstance
and SymBlockCipherDecryptionInstance to SymBlockCipherModeDecryptionInstance to
reflect this change. The API remains mostly unchanged, except for that the block
cipher mode is to be specified at instantiation time instead of when processing
an individual request for encryption or decryption.

Take advantage of the fact that the needed underlying block cipher's direction
is known ahead of request processing time and prepare + store only and
encryption or decryption implementation instance as appropriate.

Note that cfb_mode is special due to some limitations of its implementation:
even though the block cipher is only ever used in the encryption direction for
both, encryption and decryption with CFB mode, the implementations uses the
underlying block cipher instance in the decryption direction when retrieving
the final output IV. So, for the time being, instantiate a block cipher instance
capable of serving both directions and leave a fix for this to future commits.

Signed-off-by: Nicolai Stange <nstange@suse.de>