crypto/symcipher: don't obtain block, key + IV lengths from backend
In preparation of enabling the use of BoringSSL as a configurable crypto
bckend alternative, don't query the respective RustCrypto backend
implementations for either of the block, key or IV lengths, but rather
hardcode the well-known constants directly.

Introduce some macros for that and use them from existing code as appropriate.

Signed-off-by: Nicolai Stange <nstange@suse.de>