crypto/symcipher: allocate scratch block buffers on stack
Currently, a scratch block buffer is allocated on the heap by the encryption
and decryption primitives each. However, as the block size is small (16),
available as a compile-time constant by now, and zeroization would still
work reliably with this particular usage pattern, it is clearly better to just
allocate it on the stack. Do that.

Signed-off-by: Nicolai Stange <nstange@suse.de>