crypto/ecc: make CurveFieldOps private
Currently, the crypto/ecc defines a CurveOps, which contains a
CurveFieldOps. The latter handles operations on the prime field for
scalars in Montgomery form. However, the CurveFieldOps is an
implementation details and other crypto libraries like BoringSSL have no
corresponding concept in their APIs.

In preparation of enabling the use of BoringSSL as a configurable crypto
backend alternative, make CurveFieldOps private and let all exported
primitives take a CurveOps instance instead as appropriate.

Signed-off-by: Nicolai Stange <nstange@suse.de>