crypto/ecc: make {Affine,Projective}Point opaque and clarify semantics
crypto/ecc defines two point representation formats:
AffinePointMontgomeryForm and ProjectivePoint. However, users of the
API should not really be concerned with the exact internal representation,
but only care about which to use when, i.e. to choose between storage and
computational efficiency.

Moreover, other crypto libraries like BoringSSL don't expose different EC
point representations to the user at all, there's only a single opaque
EC_POINT *.

In preparation of enabling the use of BoringSSL as a configurable crypto
backend alternative, make the current semantics more clear:
- rename AffinePointMontgomeryForm to "AffinePoint"  and document
  that it /may/ have better storage efficiency,
- document that using ProjectivePoint /may/ have computational
  advantages.
Furthermore, document that the naming serves only as a hint and that the
actual internal representation is completely up to the backend
implementation and opaque to the user.

Signed-off-by: Nicolai Stange <nstange@suse.de>