crypto/ecc: rename gen_random_scalar to tcg_tpm2_gen_random_ec_scalar
The gen_random_scalar() follows the procedure outlined in the TCG TPM2
spec for generating a random scalar < the curve order and is used for
e.g. EC key generation.

However, this exact procedure needs to be followed only for key
derivations (for reproducability) and when a truly random key is to be
generated, would likely prefer to have their selected backend
implementation's respective key generation facility invoked.

In preparation of supporting that, rename "gen_random_scalar()" to
"tcg_tpm2_gen_random_ec_scalar()" for clarity and to avoid potential
confusion.

Signed-off-by: Nicolai Stange <nstange@suse.de>