bssl-bare-sys: implement support for controlling target integration
Implement support for controlling the BoringSSL build.

To simplify matters, restrict the build to compiling libcrypto
only.

In order to avoid symbol naming clashes with other copies of BoringSSL or
OpenSSL in the project, prefix all symbols.

Depend on the `bssl-bare-sys-target-integration` crate if the new Cargo
feature "target-integration" is enabled. This crate dependency is supposed
to be overriden by users for integration and control the BoringSSL build
process via the `cargo::metadata::KEY=VALUE` mechanism. Details are
to be found in the new README.md

Signed-off-by: Nicolai Stange <nstange@suse.de>